requirements.txt
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
python-dotenv==1.0.0
openai==1.6.1
chromadb==0.4.22
tiktoken==0.5.2
